backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
supabase==2.10.0
boto3==1.35.97
python-jose[cryptography]==3.3.0
httpx>=0.26,<0.28

# Testing
pytest==8.3.4
pytest-mock==3.14.0
